@@ -1,3 +1,5 @@
 module git.fractalqb.de/fractalqb/groph
 
-go 1.13
+go 1.18
+
+require golang.org/x/exp v0.0.0-20220428152302-39d4317da171
